Move subcommand code to separate file
